+++ packages/shared/src/parsers/accounts.parser.ts
@@ -1,13 +1,15 @@
+import type {z} from 'zod';
+
 import {prefixErrorResult} from '@shared/lib/errorUtils.shared';
 import {parseZodResult} from '@shared/lib/parser.shared';
 import {makeErrorResult, makeSuccessResult} from '@shared/lib/results.shared';
 
-import type {Account, AccountFromStorage, AccountId} from '@shared/types/accounts.types';
+import type {Account, AccountId} from '@shared/types/accounts.types';
+import type {EmailAddress} from '@shared/types/emails.types';
 import type {Result} from '@shared/types/results.types';
-import type {EmailAddress} from '@shared/types/utils.types';
-import {EmailAddressSchema} from '@shared/types/utils.types';
 
 import {AccountIdSchema, AccountSchema} from '@shared/schemas/accounts.schema';
+import {EmailAddressSchema} from '@shared/schemas/emails.schema';
 
 /**
  * Parses a {@link AccountId} from a plain string. Returns an `ErrorResult` if the string is not valid.
@@ -59,7 +61,7 @@ export function parseEmailAddress(maybeEmail: string): Result<EmailAddress> {
  * Converts an {@link Account} to an {@link AccountFromStorage} object that can be persisted to
  * Firestore.
  */
-export function toStorageAccount(account: Account): AccountFromStorage {
+export function toStorageAccount(account: Account): z.infer<typeof AccountSchema> {
   return {
     accountId: account.accountId,
     email: account.email,
