+++ packages/shared/src/schemas/emails.schema.ts
@@ -1,0 +1,4 @@
+import {z} from 'zod';
+
+/** Zod schema for an {@link EmailAddress}. */
+export const EmailAddressSchema = z.string().email();
