+++ packages/shared/src/types/emails.types.ts
@@ -1,0 +1,2 @@
+/** Strongly-typed type for an {@link EmailAddress}. Prefer this over plain strings. */
+export type EmailAddress = string & {readonly __brand: 'EmailAddressBrand'};
