+++ packages/sharedClient/src/services/auth.client.ts
@@ -14,8 +14,9 @@ import {
 import {asyncTry, prefixError} from '@shared/lib/errorUtils.shared';
 
 import type {AuthStateChangedCallback} from '@shared/types/accounts.types';
+import type {EmailAddress} from '@shared/types/emails.types';
 import type {AsyncResult} from '@shared/types/results.types';
-import type {Consumer, EmailAddress, Task} from '@shared/types/utils.types';
+import type {Consumer, Task} from '@shared/types/utils.types';
 
 import {firebaseService} from '@sharedClient/services/firebase.client';
