+++ packages/sharedClient/src/types/accounts.client.types.ts
@@ -5,8 +5,8 @@ import {makeErrorResult, makeSuccessResult} from '@shared/lib/results.shared';
 import {parseAccountId, parseEmailAddress} from '@shared/parsers/accounts.parser';
 
 import type {AccountId} from '@shared/types/accounts.types';
+import type {EmailAddress} from '@shared/types/emails.types';
 import type {Result} from '@shared/types/results.types';
-import type {EmailAddress} from '@shared/types/utils.types';
 
 /**
  * A generic type representing the user who is currently logged in.
